ref(revisionCopy): move copy from service to core and yml to xml
