+++ src/Core/Revision/Copy/CopyContextFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EMS\CoreBundle\Service\Revision\Copy;
+namespace EMS\CoreBundle\Core\Revision\Copy;
 
 use EMS\CommonBundle\Service\ElasticaService;
 use EMS\CoreBundle\Entity\Environment;
